Update LazySodium to version 3.5.2.

+++ crypto-v2-libsodium/src/main/java/net/aholbrook/paseto/crypto/v2/libsodium/LibSodiumV2CryptoProvider.java
@@ -50,7 +50,7 @@ public class LibSodiumV2CryptoProvider extends V2CryptoProvider {
 	public boolean ed25519Sign(byte[] sig, byte[] m, byte[] sk) {
 		validateEd25519Sign(sig, m, sk);
 
-		long[] sigLen = new long[] {sig.length};
+		int[] sigLen = new int[] {sig.length};
 		return sodium.cryptoSignDetached(sig, sigLen, m, m.length, sk);
 	}
